Update tests after RC coeff calculation refactor
One important thing to note is that the `advected_interp_method` for
`INSFVMassAdvection` absolutely matters because most of the time given
the structure of my input files, it's going to be the one that computes
the RC coefficient. And the `advected_interp_method` plays a role in the
calculation of the RC coefficient. We definitely need to create an
action to avoid this possible pitfall