Fix coupled cell-grad/ebf-value calculation for RZ
Before I was adding the RZ coordinate correction to the gradient *after*
we solved the cell-grad/ebf-value system. There is a fundamental problem
with that, which can be recognized by thinking about a simple 1D problem
where the gradient should be zero and consequently all the face values
should be the same. If you do not correct the gradient equation to
include the -phi/r term, then by definition the face value at larger r
*must* be smaller than the face value at lower r such that the summation
of Sf*phi_f equals zero for "zero gradient".