Revert "Create FaceArg::isConsumerBoundaryFace utility"
This reverts commit e5e9ee6a9ce215632855cb929161f5ad472e9f59.

At this time, I don't see a use for this in our code base. There may be
a use for it in the future, but for now the change to `isExtrapolatedBoundaryFace`
to return a pair is actually what we need in the current code base. So
what I thought was a need for a consumer-base utility is actually a need
for a producer-based utility. E.g. we have branches in our code right now
that are essentially asking "my functor may be defined here, it may not be
defined here, what should I do?". We are not limited by what the consumer
is asking for; we are limited by the data that the functor/producer has